Fixed a problem where the DistanceCheck actor property didn't work correctly, resulting in things unexpectedly not being drawn in Visual Mode

+++ Source/Core/VisualModes/VisualThing.cs
@@ -66,7 +66,7 @@ namespace CodeImp.DoomBuilder.VisualModes
 		// Rendering
 		private RenderPass renderpass;
 		private Matrix position;
-		private int cameradistance;
+		private double cameradistance;
 		private Color4 cagecolor;
 		protected bool sizeless; //mxd. Used to render visual things with 0 width and height
 		protected float fogfactor; //mxd
@@ -125,7 +125,7 @@ namespace CodeImp.DoomBuilder.VisualModes
 		
 		//mxd
 		internal int VertexColor { get { return vertices.Length > 0 && vertices[0].Length > 0 ? vertices[0][0].c : 0; } }
-		public int CameraDistance { get { return cameradistance; } }
+		public double CameraDistance { get { return cameradistance; } }
 		public float FogFactor { get { return fogfactor; } }
 		public Vector3f Center
 		{ 
@@ -241,7 +241,7 @@ namespace CodeImp.DoomBuilder.VisualModes
 		//mxd
 		internal void CalculateCameraDistance(Vector3D campos) 
 		{
-			cameradistance = (int)((CenterV3D - campos).GetLengthSq());
+			cameradistance = (CenterV3D - campos).GetLengthSq();
 		}
 		
 		// This is called before a device is reset (when resized or display adapter was changed)
